backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
websockets==12.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.6.1
python-dotenv==1.0.1